requirements.txt
Flask==0.10.1
requests==2.12.4
flask-restful-swagger==0.19
flask-restplus==0.9.2
watson-developer-cloud>=1.3.5
python-dotenv==0.8.2
cloudant==2.12.0
scipy>=0.5
numpy>=1.5
ibm-cos-sdk==2.4.4
ibm-cos-sdk-core==2.4.4
ibm-cos-sdk-s3transfer==2.4.4
ibm-db==3.0.1
ibm-db-sa==0.3.4
idna==2.8
iotfunctions==2.0.1